requirements.txt
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
cloudinary==1.40.0
colorgram.py==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pillow==10.2.0
prettytable==3.9.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.29
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
